type validation fixes, all type tests pass

+++ test/test-cases/python/expected/ydk/models/ydktest/_meta/_ydktest_sanity.py
@@ -852,7 +852,7 @@ _meta_table = {
                 ''',
                 'u_number32',
                 'ydktest-sanity', False),
-            _MetaInfoClassMember('u_number64', ATTRIBUTE, 'int' , None, None, 
+            _MetaInfoClassMember('u_number64', ATTRIBUTE, 'long' , None, None, 
                 [(0, 18446744073709551615L)], [], 
                 '''                64 bit uinteger value type
                 ''',
